use int for level instead of short

+++ src/main/java/de/Keyle/MyPet/chatcommands/CommandAdmin.java
@@ -260,7 +260,7 @@ public class CommandAdmin implements CommandExecutor, TabCompleter
             {
                 myPet.getSkills().reset();
 
-                short lvl = myPet.getExperience().getLevel();
+                int lvl = myPet.getExperience().getLevel();
                 MyPetSkillTree skillTree = myPet.getSkillTree();
 
                 if (skillTree != null)
