+++ src/main/java/de/Keyle/MyPet/event/MyPetLevelUpEvent.java
@@ -30,17 +30,17 @@ public class MyPetLevelUpEvent extends Event
     private static final HandlerList handlers = new HandlerList();
 
     private final MyPet myPet;
-    private final short level;
+    private final int level;
     private final boolean beQuiet;
 
-    public MyPetLevelUpEvent(MyPet myPet, short Level)
+    public MyPetLevelUpEvent(MyPet myPet, int Level)
     {
         this.myPet = myPet;
         this.level = Level;
         this.beQuiet = false;
     }
 
-    public MyPetLevelUpEvent(MyPet myPet, short level, boolean beQuiet)
+    public MyPetLevelUpEvent(MyPet myPet, int level, boolean beQuiet)
     {
         this.myPet = myPet;
         this.level = level;
@@ -62,7 +62,7 @@ public class MyPetLevelUpEvent extends Event
         return myPet;
     }
 
-    public short getLevel()
+    public int getLevel()
     {
         return level;
     }
